Vector Icon Button developing

+++ FinanceSaldo/ViewModel/InvoiceViewModel.cs
@@ -1,9 +1,14 @@
-﻿using System.Collections.ObjectModel;
+﻿using System;
+using System.Collections.ObjectModel;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
 using FinanceSaldo.Model;
 using GalaSoft.MvvmLight.Command;
 using GalaSoft.MvvmLight.Messaging;
+using SimpleDialogs;
+using SimpleDialogs.Controls;
+using SimpleDialogs.Enumerators;
 
 namespace FinanceSaldo.ViewModel
 {
@@ -25,6 +30,8 @@ namespace FinanceSaldo.ViewModel
             set => Set(ref _invoice, value);
         }
 
+        public CollectionView InvoiceView { get; private set; }
+
         private Invoice _selectedInvoice;
         public Invoice SelectedInvoice
         {
@@ -43,8 +50,72 @@ namespace FinanceSaldo.ViewModel
             set => Set(ref _expiryDaysList, value);
         }
 
-        public string FilterText { get; set; }
+        #region Filter
+        private string _filterText;
+        public string FilterText
+        {
+            get => _filterText;
+            set
+            {
+                Set(ref _filterText, value);
+                ResetFilterTextCommand.RaiseCanExecuteChanged();
+                InvoiceView.Refresh();
+            }
+        }
 
+        private DateTime _filterStartDate;
+        public DateTime FilterStartDate
+        {
+            get => _filterStartDate;
+            set
+            {
+                Set(ref _filterStartDate, value);
+                RaisePropertyChanged(nameof(FilterDateDif));
+                InvoiceView.Refresh();
+            }
+        }
+
+        private DateTime _filterEndDate = DateTime.Now.Date;
+        public DateTime FilterEndDate
+        {
+            get => _filterEndDate;
+            set
+            {
+                Set(ref _filterEndDate, value);
+                RaisePropertyChanged(nameof(FilterDateDif));
+                InvoiceView.Refresh();
+            }
+        }
+
+        public double FilterDateDif
+        {
+            get => (FilterEndDate.Date - FilterStartDate.Date).TotalDays;
+            set
+            {
+                FilterStartDate = FilterEndDate.AddDays(-value);
+                RaisePropertyChanged(nameof(FilterStartDate));
+            }
+        }
+
+        private ObservableCollection<string> _filterDateDifList = new ObservableCollection<string>() { "10", "20", "30", "40" };
+        public ObservableCollection<string> FilterDateDifList
+        {
+            get => _filterDateDifList;
+            set => Set(ref _filterDateDifList, value);
+        }
+
+        bool OnFilterInvoice(object item)
+        {
+            var invoice = (Invoice)item;
+            if (invoice.Date.Date < FilterStartDate.Date || invoice.Date.Date > FilterEndDate.Date) return false;
+
+            if (string.IsNullOrEmpty(FilterText)) return true;
+            if (invoice.Name == null) return false;
+            return invoice.Name.Contains(FilterText);
+        }
+        #endregion
+
+        #region Commands
         public RelayCommand AddCommand { get; set; }
         private void ExecuteAddCommand()
         {
@@ -66,9 +137,27 @@ namespace FinanceSaldo.ViewModel
             _dataService.UpdateCompany(Company);
         }
 
-        public RelayCommand DeleteCommand { get; set; }
-        private void ExecuteDeleteCommand()
+        public RelayCommand DialogDeleteCommand { get; set; }
+        private void ExecuteDialogDeleteCommand()
         {
+            DialogManager.ShowDialog(new AlertDialog()
+            {
+                AlertLevel = AlertLevel.Warning,
+                ButtonsStyle = DialogButtonStyle.YesNo,
+                YesButtonContent = "Да",
+                NoButtonContent = "Нет",
+                ExitDialogCommand = DeleteCommand,
+                ShowCopyToClipboardButton = false,
+                Message = $"Удалить запись {SelectedInvoice.Name}?",
+                Title = "Подтверждение удаления",
+            });
+        }
+
+        public RelayCommand<DialogResult> DeleteCommand { get; set; }
+        private void ExecuteDeleteCommand(DialogResult result)
+        {
+            DialogManager.HideVisibleDialog();
+            if (result != DialogResult.Yes) return;
             _dataService.RemoveInvoice(SelectedInvoice);
             Invoice.Remove(SelectedInvoice);
             Messenger.Default.Send(new NotificationMessage("TotalSaldoChanged"));
@@ -84,7 +173,7 @@ namespace FinanceSaldo.ViewModel
         private void ExecutePrintCommand(DataGrid dataGrid)
         {
             PrintDialog printdlg = new PrintDialog();
-            if ((bool) printdlg.ShowDialog().GetValueOrDefault())
+            if ((bool)printdlg.ShowDialog().GetValueOrDefault())
             {
                 Size pageSize = new Size(printdlg.PrintableAreaWidth, printdlg.PrintableAreaHeight);
                 dataGrid.Measure(pageSize);
@@ -93,30 +182,34 @@ namespace FinanceSaldo.ViewModel
             }
         }
 
-        public RelayCommand SearchCommand { get; set; }
-        private void ExecuteSearchCommand()
+        public RelayCommand ResetFilterTextCommand { get; set; }
+        private void ExecuteResetFilterTextCommand()
         {
-            foreach (Invoice inv in Invoice)
-            {
-                if (inv.Name != null && inv.Name.ToLower().Contains(SearchText))
-                {
-                    SelectedInvoice = inv;
-                }
-            }
+            FilterText = String.Empty;
         }
+        private bool CanExecuteResetFilterTextCommand()
+        {
+            return !string.IsNullOrEmpty(FilterText);
+        }
+
+        #endregion
 
         public InvoiceViewModel(IDataService dataService, Company company) : base(company.Name)
         {
             _dataService = dataService;
             Company = company;
             Invoice = company.Invoice;
+            InvoiceView = (CollectionView)CollectionViewSource.GetDefaultView(Invoice);
+            InvoiceView.Filter = OnFilterInvoice;
+            FilterDateDif = 30;
 
             AddCommand = new RelayCommand(ExecuteAddCommand);
             SaveCommand = new RelayCommand(ExecuteSaveCommand);
-            DeleteCommand = new RelayCommand(ExecuteDeleteCommand);
+            DialogDeleteCommand = new RelayCommand(ExecuteDialogDeleteCommand);
+            DeleteCommand = new RelayCommand<DialogResult>(ExecuteDeleteCommand);
             CloseTabCommand = new RelayCommand(ExecuteCloseTabCommand);
             PrintCommand = new RelayCommand<DataGrid>(ExecutePrintCommand);
-            SearchCommand = new RelayCommand(ExecuteSearchCommand);
+            ResetFilterTextCommand = new RelayCommand(ExecuteResetFilterTextCommand, CanExecuteResetFilterTextCommand);
         }
     }
 }
